% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_correct_laser.R
\name{DeepS_correct_laser}
\alias{DeepS_correct_laser}
\title{Laser Detection Correction Function}
\usage{
DeepS_correct_laser(RESULT)
}
\arguments{
\item{RESULT}{A dataset produced by the \code{DeepS_find_laser()} function.}
}
\value{
A backup file is created in the user's \code{Documents} directory.
}
\description{
This function performs corrections on the dataset produced by \code{DeepS_find_laser()} for double green lasers on a towed benthic sledge.
}
\examples{
result_data <- DeepS_find_laser(MP4_file)
corrected_data <- correct_laser_detection(result_data)
}
