% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_darckzone.R
\name{DeepS_darckzone}
\alias{DeepS_darckzone}
\title{This function creat grid count number of pixel in dark zones"}
\usage{
DeepS_darckzone(mydir, sensibility, plot)
}
\arguments{
\item{RESULT}{Is the number of not avalaible pixel is produced by DeepS_darckzone()"}
}
\value{
GRID dir is create in picture directory with picture and summary
}
\description{
This function creat grid count number of pixel in dark zones"
}
