% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_find_laser.R
\name{DeepS_find_laser}
\alias{DeepS_find_laser}
\title{Cette fonction detect les doubles lasers verts sur une luge benthique tracté"}
\usage{
DeepS_find_laser(MP4_file, para, freq)
}
\arguments{
\item{MP4_file}{est la vide assigné MP4_file <- file.choose()}

\item{para}{est une option de parallelisation  pour optiniser le temps de calcul}

\item{freq}{est la frequence des frames utilisés (elle doit etre supperieur a 40)}
}
\value{
un fichier RESULT, est une sauvegarde dans user document
}
\description{
Cette fonction detect les doubles lasers verts sur une luge benthique tracté"
}
