% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_gridpic45.R
\name{DeepS_gridpic45}
\alias{DeepS_gridpic45}
\title{Create a 45-Knot Grid on Images}
\usage{
DeepS_gridpic45(mydir)
}
\arguments{
\item{mydir}{A character string specifying the directory where the images are stored.}
}
\value{
A new subdirectory named \code{GRID} is created within the specified directory. Each image is processed and saved in this \code{GRID} directory with the grid overlay.
}
\description{
The \code{DeepS_gridpic45()} function overlays a grid with 45 knots on each image in the specified directory. This function is useful for creating grids to characterize substrates in the images.
}
\examples{
mydir <- "path/to/your/image/directory"
DeepS_gridpic45(mydir)
}
