% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_PictoBIIGLE.R
\name{DeepS_PictoBIIGLE}
\alias{DeepS_PictoBIIGLE}
\title{This function creat txt file with all picture names of the directory"}
\usage{
DeepS_PictoBIIGLE(mydir)
}
\arguments{
\item{RESULT}{Is the txt file produced by DeepS_PictoBIIGLE()"}
}
\value{
list.filetoBIIGLE.txt is produce in directory
}
\description{
This function creat txt file with all picture names of the directory"
}
