% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdeepsledge_Vhelper.R
\name{open_help_vignette}
\alias{open_help_vignette}
\title{Ouvrir la vignette de l'aide du package}
\usage{
open_help_vignette()
}
\description{
Ouvre la vignette de l'aide du package Rdeepsledge.
}
