% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_correct_laser.R
\name{DeepS_correct_laser}
\alias{DeepS_correct_laser}
\title{Cette fonction detect les doubles lasers verts sur une luge benthique tracté"}
\usage{
DeepS_correct_laser(RESULT)
}
\arguments{
\item{RESULT}{est le jeu de données types produit par DeepS_find_laser.R}
}
\value{
un fichier RESULT, est une sauvegarde dans user document
}
\description{
Cette fonction detect les doubles lasers verts sur une luge benthique tracté"
}
