% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_correct_laser.R
\name{DeepS_correct_laser}
\alias{DeepS_correct_laser}
\title{This function detects double green lasers on a towed benthic sledge"}
\usage{
DeepS_correct_laser(RESULT)
}
\arguments{
\item{RESULT}{Is the dataset type produced by DeepS_find_laser.R.}
}
\value{
a RESULT file is a backup in the user's document
}
\description{
This function detects double green lasers on a towed benthic sledge"
}
