% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_find_laser.R
\name{DeepS_find_laser}
\alias{DeepS_find_laser}
\title{Laser Detection Function}
\usage{
DeepS_find_laser(MP4_file, para, freq)
}
\arguments{
\item{MP4_file}{The video file to analyze, e.g., assigned with \code{MP4_file <- file.choose()}.}

\item{para}{A boolean indicating whether parallelization is used to optimize computing time. Default is \code{FALSE}.}

\item{freq}{The frame frequency used for processing. A shorter frequency increases the computation time.}
}
\value{
Results are saved in the user's \code{Documents} directory as \code{laser_position_$yourvideoname$.txt} for each iteration.
}
\description{
This function detects double green lasers on a towed benthic sledge from a video file.
}
\examples{
MP4_file <- file.choose()
detect_lasers(MP4_file, para = FALSE, freq = 30)
}
