% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_gridpic45.R
\name{DeepS_gridpic45}
\alias{DeepS_gridpic45}
\title{This function creat grid on new picture to help to subsample picture"}
\usage{
DeepS_gridpic45(mydir)
}
\arguments{
\item{RESULT}{Is the new pircture is produced by DeepS_gridpic45()"}
}
\value{
GRID dir is create in picture directory
}
\description{
This function creat grid on new picture to help to subsample picture"
}
