% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepS_PictoBIIGLE.R
\name{DeepS_PictoBIIGLE}
\alias{DeepS_PictoBIIGLE}
\title{Merge Picture Names into a Text File for BIIGLE Input}
\usage{
DeepS_PictoBIIGLE(mydir)
}
\arguments{
\item{mydir}{A character string specifying the directory where the pictures are stored.}
}
\value{
A text file named \code{list.filetoBIIGLE.txt} is created in the specified directory. This file contains the names of all the pictures.
}
\description{
This function generates a text file containing the names of all pictures in the specified directory. This text file can be used as an input list for BIIGLE.
}
\examples{
mydir <- "path/to/your/image/directory"
DeepS_PictoBIIGLE(mydir)
}
